Fix : 디렉토리 이름 수정 DreamConsumer-main -> DreamConsumer
