+++ .idea/modules/demo.main.iml
@@ -1,8 +1,8 @@
 <?xml version="1.0" encoding="UTF-8"?>
 <module version="4">
   <component name="NewModuleRootManager">
-    <content url="file://$MODULE_DIR$/../../server/DreamConsumer-main/build/generated/sources/annotationProcessor/java/main">
-      <sourceFolder url="file://$MODULE_DIR$/../../server/DreamConsumer-main/build/generated/sources/annotationProcessor/java/main" isTestSource="false" generated="true" />
+    <content url="file://$MODULE_DIR$/../../server/DreamConsumer/build/generated/sources/annotationProcessor/java/main">
+      <sourceFolder url="file://$MODULE_DIR$/../../server/DreamConsumer/build/generated/sources/annotationProcessor/java/main" isTestSource="false" generated="true" />
     </content>
   </component>
 </module>
