--- server/DreamConsumer-main/appspec.yml
@@ -1,18 +1,0 @@
-version: 0.0
-os: linux
-files:
-  - source:  /
-    destination: /home/ubuntu/action
-    overwrite: yes
-
-permissions:
-  - object: /
-    pattern: "**"
-    owner: ubuntu
-    group: ubuntu
-
-hooks:
-  ApplicationStart:
-    - location: scripts/deploy.sh
-      timeout: 60
-      runas: ubuntu
