--- server/DreamConsumer-main/src/main/java/com/example/demo/user/dto/UserMyPageResponseDto.java
@@ -1,24 +1,0 @@
-package com.example.demo.user.dto;
-
-import lombok.AllArgsConstructor;
-import lombok.Getter;
-
-@Getter
-@AllArgsConstructor
-public class UserMyPageResponseDto {
-    private long userId;
-
-    private String userName;
-
-    private String email;
-
-    private String job;
-
-    private int age;
-
-    // 등급
-
-    //성공한 item 개수
-
-    // 성공한 items pagination
-}
