+++ server/DreamConsumer/src/main/java/com/example/demo/exception/BusinessLogicException.java
@@ -1,0 +1,13 @@
+package com.example.demo.exception;
+
+import lombok.Getter;
+
+public class BusinessLogicException extends RuntimeException{
+    @Getter
+    private ExceptionCode exceptionCode;
+
+    public BusinessLogicException(ExceptionCode exceptionCode) {
+        super(exceptionCode.getMessage());
+        this.exceptionCode = exceptionCode;
+    }
+}
