+++ server/DreamConsumer/src/main/java/com/example/demo/user/dto/UserResponseDto.java
@@ -1,0 +1,22 @@
+package com.example.demo.user.dto;
+
+import lombok.AllArgsConstructor;
+import lombok.Getter;
+
+@Getter
+@AllArgsConstructor
+public class UserResponseDto {
+    private long userId;
+
+    private String userName;
+
+    private String email;
+
+    private String password;
+
+    private String job;
+
+    private int age;
+
+    private boolean emailAcceptance;
+}
