+++ server/DreamConsumer/src/main/java/com/example/demo/user/mapper/UserMapper.java
@@ -1,0 +1,14 @@
+package com.example.demo.user.mapper;
+
+import com.example.demo.user.dto.UserMyPageResponseDto;
+import com.example.demo.user.dto.UserPostDto;
+import com.example.demo.user.dto.UserResponseDto;
+import com.example.demo.user.entity.User;
+import org.mapstruct.Mapper;
+
+@Mapper(componentModel = "spring")
+public interface UserMapper {
+    User UserPostDtoToUser(UserPostDto userPostDto);
+    UserResponseDto UserToUserResponseDto(User user);
+    UserMyPageResponseDto UserToUserMyPageResponseDto(User user);
+}
